requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.1
django-autocomplete-light==3.7.0.dev0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
irekua-autocomplete==0.1.2
irekua-database==0.1.2
irekua-dev-settings==0.1
irekua-filters==0.1.2
irekua-models==0.1.3
irekua-permissions==0.1.2
jsonschema==3.2.0
numpy==1.19.1
Pillow==7.2.0
psycopg2-binary==2.8.5
pycountry==20.7.3
pyrsistent==0.16.0
pytz==2020.1
selia-registration==0.1
selia-templates==0.1.2
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
timezonefinder==4.4.1

